+++ AdventOfCode2022/Day11/Parsers/MonkeyStartingItemsParser.cs
@@ -5,14 +5,14 @@ namespace AdventOfCode2022.Day11.Parsers;
 public static class MonkeyStartingItemsParser
 {
     private const string StartingItemsMarker = "Starting items:";
-    public static List<int> Parse(string startingItemsInformation)
+    public static List<long> Parse(string startingItemsInformation)
     {
         var startIndex = startingItemsInformation.IndexOf(StartingItemsMarker, StringComparison.Ordinal) + StartingItemsMarker.Length;
         var itemsString = startingItemsInformation.Substring(startIndex, startingItemsInformation.Length - startIndex);
         return itemsString
             .RemoveWhitespace()
             .Split(",")
-            .Select(item => item.ToInt())
+            .Select(item => item.ToLong())
             .ToList();
     }
 }
